+++ packages/run/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./run.ts";
 export * from "./valueparser.ts";
+export * from "./print.ts";